sepolicy: bootcontrol: Remove sys_rawio capabilities
sys_rawio is not really needed by the bootcontrol HAL so removing it.

The selinux denial that comes without it is benign and shows up due to a
capability() call done in the open function by the scsi driver, which the
bootcontrol HAL calls in order to switch the UFS boot LUN.

If the driver sees this capability missing from the process that called the
open, it just modifies the type of memory buffer it requests from the
kernel(it requests a zeroed out one). It does not cause the open to fail or
have any other side effects.

Change-Id: I8890a13fdb7af0c848e5e7b9aaaaf64c1afc53b6